requirements.txt
# Core dependencies
httpx==0.25.2
beautifulsoup4==4.12.2
sqlalchemy>=2.0.36

# Database drivers (choose one)
# For PostgreSQL:
# psycopg2-binary==2.9.9

# For MySQL:
mysql-connector-python==8.2.0
# Alternative: PyMySQL==1.1.0

# Optional but recommended
python-dotenv==1.0.0
